requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
scipy>=1.10.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4